dirbpy.py
A *tiny* web content scanner based on the original Dirbuster project. 
It's a simple script to check if the connection is established over HTTP:80, and if the requested page exists; when you don't have access to more comprehensive tools. 
You provide a wordlist that has the extended paths to test.